Added programatic verification for snowtrace.io
