+++ .env.example
@@ -1,2 +1,3 @@
 PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
 WALLET_ADDR=0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266
+SNOWTRACE_KEY=RR2IFYOUCOPYPASTANGMIPDP1D4NE1QKPZ
