+++ snowtrace/args.js
@@ -1,4 +1,9 @@
+/** args.js
+ * - Used in conjunction with the Hardhat CLI for verifying single contracts
+ * - Use this file to populate whatever constructor arguements you provide when verifying
+ */
+
 module.exports = [    
-    "0x68b8037876385bbd6bbe80babb2511b95da372c4",
+    "0xe4543c234d4b0ad6d29317cfe5feecaf398f5649",
     "0xc9a51fb9057380494262fd291aed74317332c0a2"
   ];
